hoist out replsessionproxy
